Move noop template helper
