+++ packages/react-dev-overlay/src/internal/helpers/noop-template.ts
