+++ packages/react-dev-overlay/src/internal/styles/ComponentStyles.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { styles as codeFrame } from '../components/CodeFrame/styles'
 import { styles as overlay } from '../components/Overlay/styles'
-import { noop as css } from '../noop-template'
+import { noop as css } from '../helpers/noop-template'
 
 export function ComponentStyles() {
   return (
