+++ packages/react-dev-overlay/src/internal/styles/CssReset.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { noop as css } from '../noop-template'
+import { noop as css } from '../helpers/noop-template'
 
 export function CssReset() {
   return (
